00_SOFTWARE/requirements.txt
future~=0.18.2
six~=1.16.0
tinkerforge~=2.1.29
simple_pid~=1.0.1
configparser~=5.2.0
pyserial~=3.5